Deduplicate some code
